api-backend/pyproject.toml
[project]
name = "api-backend"
version = "0.1.0"
description = "FraudSense API Backend - FastAPI server for fraud detection with ML models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.90.0",
    "vllm>=0.10.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "python-magic>=0.4.27",
]
